+++ rollup.config.js
@@ -47,6 +47,14 @@ export default {
       "styled-components": path.resolve(__dirname, "node_modules", "styled-components")
     }),
     babel({
+      presets: [["@babel/preset-env", { modules: false, loose: true }], "@babel/preset-react"],
+      env: {
+        test: {
+          plugins: [["babel-plugin-styled-components", "transform-runtime"]],
+          presets: [["react-app"]]
+        }
+      },
+      plugins: [["babel-plugin-styled-components", "transform-runtime"]],
       exclude: "node_modules/**"
     }),
     commonjs()